+++ src/config/CLIFlags.js
@@ -2,12 +2,6 @@ import CLICommands from './CLICommands.js'
 
 class CLIFlags {
   /** @type {string[]}  **/
-  static BACKUP_FLAG_ACCEPTED_VALUES = ['-b', '--backup']
-
-  /** @type {string[]}  **/
-  static BACKUP_FLAG_COMMANDS = [CLICommands.INSTALL_COMMAND_NAME]
-
-  /** @type {string[]}  **/
   static WATCH_FLAG_ACCEPTED_VALUES = ['-w', '--watch']
 
   /** @type {string[]}  **/